Rewrite WiredProgressBar reusable component
