+++ src/modules/UI/components/WiredProgressBar/WiredProgressBarStyle.js
