--- src/modules/UI/scenes/WalletList/components/WalletListProgressBar/WalletListProgressBarConnector.js
@@ -1,20 +1,0 @@
-// @flow
-
-import { connect } from 'react-redux'
-
-import type { Dispatch, State } from '../../../../../ReduxTypes'
-import { getWalletLoadingPercent } from '../../../../selectors.js'
-import type { WalletListProgressBarDispatchProps, WalletListProgressBarStateProps } from './WalletListProgressBar.ui.js'
-import { WalletListProgressBarComponent } from './WalletListProgressBar.ui.js'
-
-const mapStateToProps = (state: State): WalletListProgressBarStateProps => {
-  const progressPercentage = getWalletLoadingPercent(state)
-  return {
-    progressPercentage
-  }
-}
-const mapDispatchToProps = (dispatch: Dispatch): WalletListProgressBarDispatchProps => {
-  return {}
-}
-
-export const WalletListProgressBarConnector = connect(mapStateToProps, mapDispatchToProps)(WalletListProgressBarComponent)
